packages/common/pyproject.toml
[project]
name = "common"
version = "0.1.0"
description = "ZeroTheft shared Python platform SDK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",

    # OpenTelemetry
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.45b0",
    "opentelemetry-instrumentation-redis>=0.45b0",
    "opentelemetry-instrumentation-httpx>=0.45b0",
    "opentelemetry-instrumentation-celery>=0.45b0",
    "opentelemetry-instrumentation-logging>=0.45b0",
    "opentelemetry-semantic-conventions>=0.45b0",
    "opentelemetry-exporter-otlp-proto-http>=1.24.0",

    # RLS Middleware
    "fastapi>=0.111.0",
    "starlette>=0.37.0",
    "sqlalchemy>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
aws      = ["boto3>=1.28.0"]
redis    = ["redis>=7.1.0"]
rabbitmq = ["pika>=1.3.2"]
messaging = ["aio-pika>=9.0.0"]
all      = ["boto3>=1.28.0", "redis>=7.1.0", "pika>=1.3.2", "aio-pika>=9.0.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy-cockroachdb>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/common"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/common"]
omit = ["tests/*"]
